requirements.txt
python-telegram-bot[job-queue]==21.4
gspread==5.12.0
google-auth==2.23.4

